Update forgotten remaining paths.
