+++ open_spiel/bots/CMakeLists.txt
@@ -1,3 +1,6 @@
 if (BUILD_WITH_ROSHAMBO)
   add_subdirectory(roshambo)
 endif()
+if (BUILD_WITH_XINXIN)
+  add_subdirectory(xinxin)
+endif()
