+++ open_spiel/bots/xinxin/.gitignore
@@ -1,0 +1,2 @@
+# This directory is fetched during install, do not include in OpenSpiel.
+hearts/
