+++ open_spiel/bots/xinxin/CMakeLists.txt
