+++ open_spiel/bots/xinxin/README.md
