+++ open_spiel/bots/xinxin/xinxin_bot_test.py
