+++ open_spiel/bots/xinxin/xinxin_game_generator.cc
@@ -34,9 +34,9 @@
 #include "open_spiel/abseil-cpp/absl/random/distributions.h"
 #include "open_spiel/abseil-cpp/absl/time/clock.h"
 #include "open_spiel/abseil-cpp/absl/time/time.h"
-#include "open_spiel/games/hearts/hearts/Hearts.h"
+#include "open_spiel/bots/xinxin/hearts/Hearts.h"
 #include "open_spiel/algorithms/evaluate_bots.h"
-#include "open_spiel/games/hearts/xinxin_bot.h"
+#include "open_spiel/bots/xinxin/xinxin_bot.h"
 #include "open_spiel/spiel.h"
 #include "open_spiel/spiel_bots.h"
