+++ open_spiel/bots/xinxin/xinxin_pybind11.cc
@@ -12,13 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef OPEN_SPIEL_GAMES_HEARTS_XINXIN_PYBIND11_H_
-#define OPEN_SPIEL_GAMES_HEARTS_XINXIN_PYBIND11_H_
-
+#include "open_spiel/bots/xinxin/xinxin_bot.h"
+#include "open_spiel/spiel_bots.h"
+#include "pybind11/include/pybind11/functional.h"
+#include "pybind11/include/pybind11/numpy.h"
+#include "pybind11/include/pybind11/operators.h"
 #include "pybind11/include/pybind11/pybind11.h"
+#include "pybind11/include/pybind11/stl.h"
 
 namespace open_spiel {
-void init_pyspiel_xinxin(::pybind11::module& m);
-}  // namespace open_spiel
 
-#endif  // OPEN_SPIEL_GAMES_HEARTS_XINXIN_PYBIND11_H_
+namespace py = ::pybind11;
+
+void init_pyspiel_xinxin(::pybind11::module& m) {
+  m.def("make_xinxin_bot", open_spiel::hearts::MakeXinxinBot, py::arg("params"),
+        py::arg("uct_num_runs") = 50, py::arg("uct_c_val") = 0.4,
+        py::arg("iimc_num_worlds") = 20, py::arg("use_threads") = true,
+        "Make the XinXin bot.");
+}
+
+}  // namespace open_spiel
