+++ open_spiel/games/CMakeLists.txt
@@ -125,9 +125,6 @@ endif()
 if (${BUILD_WITH_ACPC})
   add_subdirectory(universal_poker)
 endif()
-if (${BUILD_WITH_XINXIN})
-  add_subdirectory(hearts)
-endif()
 if (${BUILD_WITH_GAMUT})
   add_subdirectory(gamut)
 endif()
